ccmp2: rework env_get_location()
This function returns the location of the environment basing on the
selected boot media. If the boot media (as selected by strapping) is not
one of the configured ones, it defaults to using no stored environment
(which will use the default one only in RAM).

On the CCMP2x, we want the SOM to use the default media (eMMC) by default
no matter the selected boot media.

https://onedigi.atlassian.net/browse/DEL-8790

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>