bootcount: ccmp25: enable bootcount mechanism to use 'User RAM' from external RTC
Bootcount support is required for dualboot mechanism, so this commit integrates
this feature to store the boot count value in the 'User RAM' registers of the
development board RTC device.

The external RTC device used in the development board (Micro Crystal RV-3028)
provides 2 general purpose 'User RAM' registers to store data and backup them
with battery. These registers are also protected against soft resets, so they
are the perfect place to store values between reboots. The address of these
registers in the RTC chip are 0x1F and 0x20.

Also this commit enables the 'bootcount' command to be able to read and reset
the 'bootcount' value from U-Boot.

https://onedigi.atlassian.net/browse/DUB-1086

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>