ccmp25-dvk: configure RTC's backup switching mode
This is in the context of using the RTC to save the boot counter.

In the ccmp25-dvk, the RTC loses power on every reset (software or hardware),
and thus, the boot counter is lost. The RTC relies on a battery backup power
domain (VBACKUP), connected to a supercap or a coin cell in the DVK, but that
requires modify the default RTC configuration in the internal EEPROM.
By default the backup switching mode (BSM) is set to zero (which means
switchover is disabled), this commit sets the BSM to level switching mode (LSM).

https://onedigi.atlassian.net/browse/DUB-1086

Signed-off-by: Javier Viguera <javier.viguera@digi.com>
(backport from commit b8c352b21cb50074ce5f42d4c7b9925d5d29e20a)
Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>