tools: digi: make_imxboot_ccimx8x: support SBC express
Extend the "make_imxboot_ccimx8x.sh" script to support building bootloaders
for SBC Express. To prevent code duplication, use a symlink and check
the name at runtime to guess if we are building an image for the SBC pro
or express.

Also, enable the new supported platforms on the Jenkins build script.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>