ccimx8x: increase memory available for dynamic allocation
As we did not have a defined value for this platform, u-boot was built
with the default Kconfig value (0x400000). That value is not big enough
to provide memory for programming sparse images using fastboot after the
default chunk size got increased for better performance in this commit:

https://source.denx.de/u-boot/u-boot/-/commit/b1aade87ca1f5d3788bd6ef570b438c2dc7a9428

Use the same malloc values as the imx8qxp-mek reference board.

https://onedigi.atlassian.net/browse/DUB-1091

Signed-off-by: Javier Viguera <javier.viguera@digi.com>