ccimx6ul: remove buggy not used function
The overloaded env_get_offset_redund function implementation for the
ccimx6ul has an infinite recursion bug. It is recursively calling itself
with no trivial case to finish.

Looking at the rest of the code, the overloaded function is not being used,
so just remove the dead code and use the weak function in `env/env.c`
where needed.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>