trustfence: move common code to hab.c and ahab.c
Move code which is common for HAB devices to hab.c, and code which is
common for AHAB devices to ahab.c, and build both unconditionally of the
platform provided they have configs CONFIG_IMX_HAB or CONFIG_AHAB_BOOT
respectively.

Declare weak functions for those implemented in the board code.

https://onedigi.atlassian.net/browse/DEL-9399

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>