trustfence: move dek copy to 'digi_auth_image'
Instead of reading a DEK copy into RAM during initialization, read it
immediately before it is going to be used, during the authentication
of Linux artifacts.

Verify first the fixed RAM address does not contain a DEK with a valid
HAB header to avoid reading the DEK blob multiple times during each
artifact authentication.

Use a fixed RAM address that does not depend on the "loadaddr"
environment variable to match the fixed address used during Kernel
artifacts signing and encryption.

https://onedigi.atlassian.net/browse/DEL-9399

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>