trustfence: ahab: fix next container alignment
AHAB containers are stored consecutively inside the imx-boot image,
but they maintain a 1K alignment.

Fix the way we look for the next container so we round up the next
container relative offset instead of rounding up the next container
absolute address, which depended on the imx-boot load address to be
aligned to 1K to work.

Also, explicitly use CONTAINER_HDR_ALIGNMENT, which is 1K, to make it
clearer.

https://onedigi.atlassian.net/browse/DEL-9399

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>