trustfence: ccimx9: don't report encryption status when booting from USB
AHAB devices report the U-Boot is encrypted after reding the eMMC
content and looking for a DEK blob. If you are booting signed images
from USB, the MMC could contain an encrypted U-Boot or not, but it has
nothing to do with the encryption status of the running U-Boot.

Ignore the MMC content when booting from USB and just don't report
anything, like we do on CC8X.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>