ccimx8x: fix ram_size computing on 4GB non-SPL build
4GB RAM is split between 2 memory banks.

The size of both SDRAM banks was being computed only for builds with
SPL.
Extend that function to compute the size of both banks for builds
without SPL too.

Also, ensure the PHYS_SDRAM_1_SIZE macro can fit a 64-bit value so we
can define by default a 4GB size, which requires 33 bits.

https://onedigi.atlassian.net/browse/DEL-9435

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>