ccimx8/ccimx9: remove '_ov_' prefix from Digi DT overlays
Device tree overlays now have the extension 'dtso' that
distinguishes them from board 'dts' files, so there is
no need for a prefix '_ov_' to tell if a file contains
a DT overlay.

To make them shorter and easier to tell the platform they
are for, change the filename format to:

  <platform-name>_<functionality>.dtso

where <platform-name> can be the name of the SOM or the
name of the DVK, so there is no need either to specify
'som' or 'board' on the filename.

These are only used to set the variable 'som_overlays'
exclusively used by Android.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>