ccimx6ul: increase partitions for 256 MiB variant
The default partition table for the 256 MiB variant had
- linux (12 MiB)
- recovery (14 MiB)
- rootfs (122 MiB)
- update (103 MiB)

DEY-5.0 default Qt images don't fit in said sizes on 'recovery'
and 'rootfs' partitions.
Considering that the SWU image (Qt) does not fit on the update
partition anyway, this commit enlarges 'linux', 'recovery' and
'rootfs', and adds also a 'data' partition at the expense of
'update', as follows:

- linux (16 MiB)
- recovery (20 MiB)
- rootfs (140 MiB)
- update (70 MiB)
- data (5 MiB)

Similar adjustments have been made to add 'data' partition or
volume on other variants layouts (and dual-boot layouts), and
the UBI volumes for single-mtd schema.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DEL-9514